영화 및 TV 상세 화면 수정

+++ core/model/src/main/java/com/bowoon/model/CombineCredits.kt
@@ -32,7 +32,7 @@ data class CombineCreditsCast(
     val releaseDate: String? = null,
     val title: String? = null,
     val video: Boolean? = null,
-    val voteAverage: Double? = null,
+    val voteAverage: Float? = null,
     val voteCount: Int? = null
 )
 
@@ -59,7 +59,7 @@ data class CombineCreditsCrew(
     val releaseDate: String? = null,
     val title: String? = null,
     val video: Boolean? = null,
-    val voteAverage: Double? = null,
+    val voteAverage: Float? = null,
     val voteCount: Int? = null
 )
