+++ core/network/src/main/java/com/bowoon/network/model/NetworkTMDBCombineCredits.kt
@@ -62,7 +62,7 @@ data class NetworkTMDBCombineCreditsCast(
     @SerialName("video")
     val video: Boolean? = null,
     @SerialName("vote_average")
-    val voteAverage: Double? = null,
+    val voteAverage: Float? = null,
     @SerialName("vote_count")
     val voteCount: Int? = null
 )
@@ -112,7 +112,7 @@ data class NetworkTMDBCombineCreditsCrew(
     @SerialName("video")
     val video: Boolean? = null,
     @SerialName("vote_average")
-    val voteAverage: Double? = null,
+    val voteAverage: Float? = null,
     @SerialName("vote_count")
     val voteCount: Int? = null
 )
