+++ core/network/src/main/java/com/bowoon/network/model/NetworkTMDBTrendingMovie.kt
@@ -46,7 +46,7 @@ data class NetworkTMDBTrendingMovieResult(
     @SerialName("video")
     val video: Boolean? = null,
     @SerialName("vote_average")
-    val voteAverage: Double? = null,
+    val voteAverage: Float? = null,
     @SerialName("vote_count")
     val voteCount: Int? = null
 )
