+++ core/network/src/main/java/com/bowoon/network/model/NetworkTMDBTrendingPeople.kt
@@ -80,7 +80,7 @@ data class NetworkTMDBTrendingPeopleKnownFor(
     @SerialName("video")
     val video: Boolean? = null,
     @SerialName("vote_average")
-    val voteAverage: Double? = null,
+    val voteAverage: Float? = null,
     @SerialName("vote_count")
     val voteCount: Int? = null
 )
